LayoutTests/imported/w3c/web-platform-tests/dom/events/mouse-event-retarget-expected.txt

Hello

PASS offsetX is correctly adjusted
